pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "my-config"
version = "0.1.0"
description = "Configuration management for Python applications"
authors = [{name = "Your Name", email = "your.email@example.com"}]
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
    "pytz",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black",
    "isort",
]